refactor: split tests
