+++ src/dotnet-affected.Tests/PathGenerationTests.cs
@@ -8,9 +8,9 @@ using Xunit;
 
 namespace Affected.Cli.Tests
 {
-    public class RepositoryPathTests
-    {
-        private class TestPathsClassData : IEnumerable<object[]>
+    public class PathGenerationTests
+    {        
+        private class RepositoryPathsClassData : IEnumerable<object[]>
         {
             public IEnumerator<object[]> GetEnumerator()
             {
@@ -37,7 +37,7 @@ namespace Affected.Cli.Tests
         }
 
         [Theory]
-        [ClassData(typeof(TestPathsClassData))]
+        [ClassData(typeof(RepositoryPathsClassData))]
         public void Should_determine_repository_path_correctly(
             string repositoryPath,
             string solutionPath,
@@ -57,5 +57,50 @@ namespace Affected.Cli.Tests
 
             Assert.Equal(expected, data.RepositoryPath);
         }
+        
+        private class OutputDirPathsClassData : IEnumerable<object[]>
+        {
+            public IEnumerator<object[]> GetEnumerator()
+            {
+                yield return new object[]
+                {
+                    "/home/lchaia/dotnet-affected", "", "/home/lchaia/dotnet-affected"
+                };
+                yield return new object[]
+                {
+                    "/home/lchaia/dotnet-affected", "relative/path",
+                    Path.Combine("/home/lchaia/dotnet-affected", "relative/path")
+                };
+                yield return new object[]
+                {
+                    "/home/lchaia/dotnet-affected", "/some/absolute/path", "/some/absolute/path"
+                };
+            }
+
+            IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+        }
+
+        [Theory]
+        [ClassData(typeof(OutputDirPathsClassData))]
+        public void Should_determine_output_dir_correctly(
+            string repositoryPath,
+            string outputDir,
+            string expected)
+        {
+            var data = new CommandExecutionData(
+                repositoryPath,
+                String.Empty,
+                string.Empty,
+                string.Empty,
+                false,
+                Enumerable.Empty<string>(),
+                new string[0],
+                false,
+                outputDir,
+                string.Empty);
+
+            Assert.Equal(expected, data.OutputDir);
+        }
+
     }
 }
